+++ src/relax/op/tensor/manipulate.h
@@ -25,6 +25,7 @@
 #define TVM_RELAX_OP_TENSOR_MANIPULATE_H_
 
 #include <tvm/relax/attrs/manipulate.h>
+#include <tvm/runtime/container/variant.h>
 
 #include "../op_common.h"
 #include "tvm/relax/expr.h"
@@ -43,7 +44,7 @@ Expr broadcast_to(Expr x, Expr shape);
  * If it is `NullOpt`, the input tensor is required to be flattened before concatenation.
  * \return The concatenated tensor.
  */
-Expr concat(Expr tensors, Optional<Integer> axis);
+Expr concat(Expr tensors, Optional<int64_t> axis);
 
 /*!
  * \brief Insert new axes at the positions given by `axis`.
@@ -153,7 +154,7 @@ Expr collapse_sum_to(Expr data, Expr shape);
  * from the backward. By default, use the flattened input array, and return a flat output array.
  * \return The computed result.
  */
-Expr repeat(Expr data, int repeats, Optional<Integer> axis = NullOpt);
+Expr repeat(Expr data, int repeats, Optional<int64_t> axis = NullOpt);
 
 /*!
  * \brief Construct an array by repeating data the number of times given by reps.
